allow only lua source file

+++ src/run.sh
@@ -51,7 +51,11 @@ case "${sol#*.}" in
         echo 'run=(java -client -Djava.security.manager -Djava.security.policy=/build/java.policy Solution)' > run_cmd;;
 	js)
 		echo 'run=(/usr/bin/js -e "delete load;delete read;delete run;delete snarf" solution.js)' > run_cmd;;
-        lua)
+    lua)
+        if [[ "$(file solution.lua)" =~ bytecode ]]; then
+            echo "Lua bytecode not allowed."
+            exit 1
+        fi
 		luac solution.lua
 		rm solution.lua
 		echo 'run=(lua /build/luasandbox.luac)' > run_cmd;;
